Add author information to guide generator
This is part of #460. It still requires styling in devsite as well as the relevant information of authors in this repository.
In terms of configuration, we require all guides to add an object with the relevant links and image link.
Then we need a separate folder with all profile images of the authors, or a generic one if they don't want to upload that.
For now, we ignore all author configuration that is just a string.

An alternative would be to have a separate configuration file where we store this information and then keep the author
configuration as a string, which is an index of that configuration. If we have several guides of the same author,
this would probably save us some copy-pasting. However, I am punting that to after the styling, to keep PRs easy.