"Identify resources loaded from the network" cleanup (#565)
Fixes #223.

Addresses (most) of the feedback there from @alankent.

There are some style-related comments that I'm not sure need to be fixed, e.g. the use of contractions seems okay. The reference to "dark forces" affecting the network wasn't my wording to begin with (maybe it was Meggin?) so I'm hesitant to change it.