Update documentation for `<a>` tag
If clicking on an element adds new content in the web page, we should still use `<button>` instead. As native `<a>` has properties like opening the link in new window or tab , this feature won't be helpful for loading the new content unless we change the url of the page.