Tweak example for feature detection and using lazysizes
Using a custom data attribute instead of a CSS class, teaches young developers to avoid CSS classes for things they were never meant to be used for.  
If a young developer starts using classes as hooks for their JavaScript code, some day in the future some other young developer will *optimize* the HTML and remove that *unused* CSS class that is nowhere to be found in any stylesheet.