fix(SnackbarContainer): check`localStorage` before intercepting install prompt, fixes #2023
The `if (!this.acceptedCookies)` would always be `true` on initial start of the app because the initial state didn't use the `localStorage` value. So we trigger a check if the user has accepted the cookies, which updates the store before we get the store's state in the component and run the check.

Now you will be greeted with the prompt after you accept cookies (which we could stop), and every subsequent visit should behave as normal. This only impacts the user until they accept cookies.