Update index.md
Suggest rm this section because it's misleading.

This doesn't seem to be the case any more. While I see LCP on my timeline, I cannot find any such marks or measures in Performance.entries(). The marks I can find are in ``.

I do see some `paint` entries on `performance.getEntriesByType('paint')`, but not the largest-contentful-paint.

I would update with the new/correct way to get this measure, but I can't figure out how to see it, maybe it's no longer gettable?