Minor tweaks to index.md
Good post!

I've done these suggestions as a PR, but you may want to ignore all or part of this. Quite fiddly to do this as individual (multiline) suggestions.

BTW: ` The contents of this method should aim to get access to the stream source` sounds odd to me. I've tweaked the text, but not sure if I've got it quite right.

BTW 2: love the use of `1_000`, `10_000`, etc.!