chore: make changes to text formatting
Found out there was an uneven text format pattern in terms of how the content is shown on the footer. 
Only the Community Guidelines, Terms & Privacy, and Web Fundamentals capitalized the first letter of each word while the rest that has two words which makes them irregular on the navbar.
So I decided to capitalize the first letters of text with two words under one to make it even among the readers.

I also added a soothing phrase to replace the developer.chrome.com text included formerly so it can be more readble among people viewing the site.