Add flags to the ffmpeg command for creating GIFs
This adds flags to the ffmpeg command for creating an MP4 from a GIF that tag the output in the sYCC color space, which fixes the "washed out" colors and color shifts often reported when doing this conversion. Without the flags, browsers tend to interpret the colorspace incorrectly.

Note this currently does not affect Firefox, but does improve display in Chrome and Safari.