Remove potential reproduction bug
I tried the code and it was super helpful but it seems like you are attempting to start scrolling from the border or outside the frame. 

The change makes it so you start scrolling from the bottom of the page. 

It gave me weird but where the whole page tried to move up. Like lifting a curtain on iPhone.