Learn Accessibility course: update to WCAG 2.2 release date (#9305)
* Learn accessibility course - color contrast module

I noticed an oversight about font sizes and ratios in the color contrast module. This should help fill in some gaps related to the upcoming testing module.

* line breaks

* line break and link

* Learn Accessibility course: WCAG 2.2 release date

WCAG 2.2 will not be released in late 2022, but the new date is sometime in 2023. I purposefully did not add the phrase "early 2023," as it is not yet clear when it will actually be released.

Co-authored-by: Alexandra White <alexandrawhite@google.com>