`/import-maps-in-all-modern-browsers/`: Fix examples by replacing `lodash` with `d3-array`. (...)
The snippets in this article don't work together because the URL for `lodash`
used points to a file that's not a module and contains no exports. Simply
replacing `lodash` with `lodash-es` causes the example to work, but `lodash`'s
main file transitively imports over 600 internal modules - which seems like a
bad example to set. Even importing `lodash-es/map.js` directly still
transitively imports over 100 modules. To both fix the example and avoid loading
an unnecessarily large dependency tree for such a simple example, I've replaced
`lodash` with `d3-array`, which is a module and has no transitive imports.