Update index.md: Replaced the term "inheritable" with "inherited by default"
This page uses the term "inheritable" which is both potentially misleading and not used in the CSS spec. This is potentially misleading because the word "inheritable" can be interpreted as "able to be inherited" as opposed to how it is being used here, which is "inherited by default". Additionally the CSS spec does not use the word "inheritable" instead referring to properties that inherit their values by default as "inherited properties".

Everywhere the word "inheritable" is used, I have replaced it with "inherited by default", which is how the word seems to have been used on this page.